experimenting with electron recorder integration
