+++ example/movie.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+  <title>Test movie</title>
+  <head>
+    <meta charset="UTF-8">
+    <link rel="stylesheet" type="text/css" href="main.css">
+  </head>
+  <body>
+    <script src="movie-bundle.js" /></script>
+  </body>
+</html>
